fix: improve firebase test account deletion process

+++ packages/near-fast-auth-signer-e2e-tests/utils/firebase.ts
@@ -51,7 +51,7 @@ const addAccountPublicKeyToFirestore = async (accountId: string, publicKey: stri
 export const deleteUserByEmail = async (email: string) => {
   try {
     const user = await admin.auth().getUserByEmail(email);
-    if (user) {
+    if (user && user.uid) {
       await deleteAccount(user.uid);
     }
   } catch (error) {
